Client now supports piped io for server data
The data that the client retrieves from the server is now sent back to
the main bluefire thread via a PipedIO reader/writer. This allows the
actual reading of the data from the server to be done much faster in the
case of large amounts of data. It also means that the client does not
allocated massive arrays of data to store all of the data from the
server.

The error chain has now been made a global variable (EWW). This is
probably temporary, but it was done to allow multiple threads from
different types to centralize their errors in one place within the same
'main' package.