Stabilized multithreaded backend and connections The multithreaded backend now performs several more tasks asynchronosly and is more stabilized. More work needs to be done to enhance the error stream that is held in the backend and determine the efficacy of keeping the error stream. Currently; as many errors that can be handled syncrhonously are handled as such to increase the user's ability to interact with the frontend.
Furthermore, connection handling has been stabilized and rewritten to
make more sense and do a better job of performing asynchronously to give
the frontend a more quick feel and reponse time.

Please note that DLE has not been enabled in the client portion of the
project yet because of certian limitations stemming from the Bluetooth
Low Energy library that have yet to be solved.

A large response time and speed increase is expected upon implementing
DLE.