Cleaned up locking, and logic around indications THe server no longer stops and restarts the shell it holds. This was added complexity that didn't really exist for a good reason.
Locking is now bi-dependant between the Bluetooth Read function and the
thread that serves as our data bufferer. This has the effect of ensuring
that the Bluetooth Read function recieves the lock after the buffer
thread has determined that there is data to be read. This has the added
benefit of speeding up data transmission because there should not be any
more data races concerning this mechanism, and there is no lock
thrashing.

When the buffering thread determines that there is data to be read,
execution turns very synchronised.